% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_headshots}
\alias{load_headshots}
\title{Load Player ID Map With Headshot Info}
\usage{
load_headshots()
}
\value{
A tibble of player names and ids from various sources.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{load_headshots()} has been deprecated due to lack of support. Please use \code{baseballr::chadwick_player_lu()}
}
\examples{
\donttest{
# load_headshots()
}

}
