% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_milb_teams}
\alias{load_milb_teams}
\title{Load MiLB Team Colors, and Logos}
\usage{
load_milb_teams()
}
\value{
A tibble of team-level abbreviations, image URLs, and league info for Minor League Baseball Teams.
}
\description{
Loads team information and logos - useful for plots!
}
\examples{
\donttest{
load_milb_teams()
}

}
\seealso{
Issues with this data should be filed here: \url{https://github.com/camdenk/mlbplotR}
}
