% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_mlb_teams}
\alias{load_mlb_teams}
\title{Load MLB Team Colors, and Logos}
\usage{
load_mlb_teams()
}
\value{
A tibble of team-level abbreviations, image URLs, and hex color codes.
}
\description{
Loads team colors, and logos - useful for plots!
}
\examples{
\donttest{
load_mlb_teams()
}

}
\seealso{
Issues with this data should be filed here: \url{https://github.com/camdenk/mlbplotR}
}
