% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mlb.R
\name{theme_mlb}
\alias{theme_x_mlb}
\alias{theme_y_mlb}
\title{Theme for MLB Team Logos}
\usage{
theme_x_mlb()

theme_y_mlb()
}
\value{
A ggplot2 theme created with [ggplot2::theme()].
}
\description{
These functions are convenience wrappers around a theme call
  that activates markdown in x-axis and y-axis labels
  using [`ggtext::element_markdown()`].
}
\details{
These functions are a wrapper around the function calls
  `ggplot2::theme(axis.text.x = ggtext::element_markdown())` as well as
  `ggplot2::theme(axis.text.y = ggtext::element_markdown())`.
  They are made to be used in conjunction with [`scale_x_mlb()`] and
  [`scale_y_mlb()`] respectively.
}
\examples{
\donttest{
library(mlbplotR)
library(ggplot2)

team_abbr <- valid_team_names()
# remove conference logos from this example
team_abbr <- team_abbr[!team_abbr \%in\% c("AL", "NL", "MLB")]

df <- data.frame(
  random_value = runif(length(team_abbr), 0, 1),
  teams = team_abbr
)
if (utils::packageVersion("gridtext") > "0.1.4"){
  ggplot(df, aes(x = teams, y = random_value)) +
    geom_col(aes(color = teams, fill = teams), width = 0.5) +
    scale_color_mlb(type = "secondary") +
    scale_fill_mlb(alpha = 0.4) +
    scale_x_mlb() +
    theme_minimal() +
    # theme_*_mlb requires gridtext version > 0.1.4
    theme_x_mlb()
}
}
}
\seealso{
[`theme_x_mlb()`], [`theme_y_mlb()`]
}
