% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_headshots}
\alias{load_headshots}
\title{Output MLB Team Abbreviations}
\usage{
load_headshots()
}
\value{
A tibble of player names and ids from various sources.
}
\description{
Output MLB Team Abbreviations
}
\examples{
\donttest{
load_headshots()
}

}
