% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cap_logos.R, R/geom_from_path.R,
%   R/geom_lines.R, R/geom_mlb_headshots.R, R/geom_mlb_logos.R,
%   R/mlbplotR-package.R
\docType{data}
\name{GeomMLBdarkcaplogo}
\alias{GeomMLBdarkcaplogo}
\alias{GeomMLBlightcaplogo}
\alias{GeomFromPath}
\alias{GeomRefLines}
\alias{GeomMLBheads}
\alias{GeomMLBlogo}
\alias{GeomMLBscoreboardlogo}
\alias{mlbplotR}
\alias{mlbplotR-package}
\title{mlbplotR: Create 'ggplot2' and 'gt' Visuals with Major League Baseball Logos}
\format{
An object of class \code{GeomMLBdarkcaplogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomMLBlightcaplogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomFromPath} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomRefLines} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.

An object of class \code{GeomMLBheads} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomMLBlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{GeomMLBscoreboardlogo} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomMLBdarkcaplogo

GeomMLBlightcaplogo

GeomFromPath

GeomRefLines

GeomMLBheads

GeomMLBlogo

GeomMLBscoreboardlogo
}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to help visualize Major League Baseball analysis in 'ggplot2' and 'gt'. You provide team/player information and 'mlbplotR' will transform that information into team colors, logos, or player headshots for graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/camdenk/mlbplotR}
  \item \url{https://camdenk.github.io/mlbplotR/}
  \item Report bugs at \url{https://github.com/camdenk/mlbplotR/issues}
}

}
\author{
\strong{Maintainer}: Camden Kay \email{camden.kay23@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Sebastian Carl \email{mrcaseb@gmail.com}
}

}
\keyword{datasets}
\keyword{internal}
